Imported another printf library
Don't feel like spending a bunch of time fiddling with it in a kernel
environment, maybe in another repo in the future.
Moved all printfs to new functions. Disabled floating point support
since we haven't enabled it yet in the kernel. Updated some formatting
and some warnings to try to cut down on extrenious warnings.