Got new bootmem allocator mostly working
It is a bitmap allocator, it is essentially a copy of the pmm allocator.
We will use it to allocate required structures for the buddy allocator
then transfer ownership of pages.

Also defined some new atomic operations. Probably shouldn't be used as
spinlocks but it works well for ref counters.

I am trying to follow a MM scheme similar to the linux ~2.6 kernel. So
the buddy allocator will be a physical memory manager, and we will have
some other allocator (tree based?) for virtual memory.